refactor: break up ConversationEventReceiver into separate event handlers
